requirements.txt
fastapi 
uvicorn 
authlib
passlib[bcrypt] 
pydantic-settings
pytest
httpx
pandas
